Split out FieldType into FieldType.{h,cpp}
