+++ UI/ClientUI.cpp
@@ -29,7 +29,7 @@
 #include "../universe/Tech.h"
 #include "../universe/Special.h"
 #include "../universe/Species.h"
-#include "../universe/Field.h"
+#include "../universe/FieldType.h"
 #include "../universe/Enums.h"
 #include "../combat/CombatLogManager.h"
 #include "../client/human/HumanClientApp.h"
