+++ msvc2017/FreeOrion/FreeOrion.vcxproj.filters
@@ -372,6 +372,9 @@
     <ClInclude Include="..\..\universe\Field.h">
       <Filter>Header Files\universe</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\universe\FieldType.h">
+      <Filter>Header Files\universe</Filter>
+    </ClInclude>
     <ClInclude Include="..\..\UI\ModeratorActionsWnd.h">
       <Filter>Header Files\UI</Filter>
     </ClInclude>
